requirement.txt
phidata
duckdb
kagglehub
streamlit
pandas
google-generativeai
